uci.c: Fix loss of precision on integer division
bench: 3161852

+++ Source/uci.c
@@ -358,8 +358,8 @@ void time_control(position_t *pos, thread_t *threads, char *line) {
       limits.time -= MIN(limits.time / 2, move_overhead);
 
       const int64_t base_time = (limits.movestogo > 0)
-          ? (limits.time / limits.movestogo) + limits.inc
-          : limits.time * DEF_TIME_MULTIPLIER + limits.inc * DEF_INC_MULTIPLIER;
+          ? (int64_t)((double)limits.time / limits.movestogo + limits.inc)
+          : (int64_t)(limits.time * DEF_TIME_MULTIPLIER + limits.inc * DEF_INC_MULTIPLIER);
 
       limits.max_time   = MAX(1, limits.time * MAX_TIME_MULTIPLIER);
       limits.hard_limit = threads->starttime + limits.max_time;
